requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
black==22.10.0
click==8.1.3
Django==4.1.3
django-phonenumber-field==7.0.0
flake8==5.0.4
mccabe==0.7.0
mypy-extensions==0.4.3
pathspec==0.10.1
phonenumbers==8.12.57
Pillow==9.3.0
platformdirs==2.5.2
pycodestyle==2.9.1
pyflakes==2.5.0
sqlparse==0.4.3
tomli==2.0.1
typing-extensions==4.4.0
